requirements.txt
psycopg2-binary
sqlalchemy
pyspark
confluent_kafka
pandas
seaborn
matplotlib
python-dotenv
kaggle
unidecode